Clippy my old pal

+++ src/ball.rs
@@ -80,7 +80,7 @@ impl Ball {
             self.time_running = 0.0;
             self.velocity_change_time_at = rng.gen_range(
                 self.random_direction_change_time_min..=self.random_direction_change_time_max,
-            ) as f32;
+            );
         }
 
         self.pos.x += self.velocity.x * delta_time;
@@ -92,7 +92,7 @@ impl Ball {
         self.pos.y -= input_axes.y * BALL_NUDGE_RATE * delta_time;
 
         // This is for bounds checking on the ball
-        // The addition or substraction inside the logic is so the circle does not use the center as
+        // The addition or subtraction inside the logic is so the circle does not use the center as
         // the x or y location. This way the circle would not go through some of the borders.
 
         if (self.pos.x + BALL_RADIUS) >= 1.0 {
